+++ .idea/modules.xml
@@ -2,8 +2,8 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
-      <module fileurl="file://$PROJECT_DIR$/telodoy.iml" filepath="$PROJECT_DIR$/telodoy.iml" />
-      <module fileurl="file://$PROJECT_DIR$/android/telodoy_android.iml" filepath="$PROJECT_DIR$/android/telodoy_android.iml" />
+      <module fileurl="file://$PROJECT_DIR$/opendrop.iml" filepath="$PROJECT_DIR$/opendrop.iml" />
+      <module fileurl="file://$PROJECT_DIR$/android/opendrop_android.iml" filepath="$PROJECT_DIR$/android/opendrop_android.iml" />
     </modules>
   </component>
 </project>
