+++ android/app/src/main/kotlin/com/example/opendrop/MainActivity.kt
@@ -1,4 +1,4 @@
-package com.example.telodoy
+package com.example.opendrop
 
 import io.flutter.embedding.android.FlutterActivity
