+++ pubspec.yaml
@@ -1,4 +1,4 @@
-name: telodoy
+name: opendrop
 description: A cross-platform file sharing app using broadcast discovery.
 version: 0.1.0
