+++ web/index.html
@@ -23,13 +23,13 @@
   <!-- iOS meta tags & icons -->
   <meta name="mobile-web-app-capable" content="yes">
   <meta name="apple-mobile-web-app-status-bar-style" content="black">
-  <meta name="apple-mobile-web-app-title" content="telodoy">
+  <meta name="apple-mobile-web-app-title" content="opendrop">
   <link rel="apple-touch-icon" href="icons/Icon-192.png">
 
   <!-- Favicon -->
   <link rel="icon" type="image/png" href="favicon.png"/>
 
-  <title>telodoy</title>
+  <title>opendrop</title>
   <link rel="manifest" href="manifest.json">
 </head>
 <body>
