+++ windows/runner/Runner.rc
@@ -90,12 +90,12 @@ BEGIN
         BLOCK "040904e4"
         BEGIN
             VALUE "CompanyName", "com.example" "\0"
-            VALUE "FileDescription", "telodoy" "\0"
+            VALUE "FileDescription", "opendrop" "\0"
             VALUE "FileVersion", VERSION_AS_STRING "\0"
-            VALUE "InternalName", "telodoy" "\0"
+            VALUE "InternalName", "opendrop" "\0"
             VALUE "LegalCopyright", "Copyright (C) 2025 com.example. All rights reserved." "\0"
-            VALUE "OriginalFilename", "telodoy.exe" "\0"
-            VALUE "ProductName", "telodoy" "\0"
+            VALUE "OriginalFilename", "opendrop.exe" "\0"
+            VALUE "ProductName", "opendrop" "\0"
             VALUE "ProductVersion", VERSION_AS_STRING "\0"
         END
     END
